+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.6
+
+- move ubcpdk.simulations  to ubcpdk.simulation for consistency with gdsfactory
+- add ubcpdk.simulation.gmeep
+- move tests to ubcpdk/tests
+
 ## 1.0.5
 
 - move sample data into the module
@@ -7,7 +13,7 @@
 ## 1.0.0
 
 - rename package from ubc to ubcpdk to match pypi name
-- move ubcsp into ubcpdk/simulations/circuits
+- move ubcsp into ubcpdk/simulation/circuits
 - rename ubcpdk/da as ubcpdk/data
 
 ## 0.0.12